+++ domains/auth/Auth.API/API/Program.cs
@@ -1,10 +1,10 @@
+using API.Configuration;
 using API.Services;
 using Serilog;
 using Serilog.Formatting.Json;
 using System.Runtime.CompilerServices;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using API.Helpers;
 
 [assembly: InternalsVisibleTo("Tests")]
