+++ domains/auth/Auth.API/API/Services/CryptographyService.cs
@@ -1,4 +1,4 @@
-using API.Helpers;
+using API.Configuration;
 using Microsoft.Extensions.Options;
 using System.Security.Cryptography;
 using System.Text;
