+++ domains/auth/Auth.API/API/Services/OidcProviders/SignaturGruppen.cs
@@ -1,7 +1,7 @@
-using API.Helpers;
+using API.Configuration;
 using API.Models;
-using System.Text.Json;
 using Microsoft.Extensions.Options;
+using System.Text.Json;
 
 namespace API.Services;
