+++ domains/auth/Auth.API/API/Services/OidcService.cs
@@ -1,4 +1,4 @@
-using API.Helpers;
+using API.Configuration;
 using API.Models;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.Extensions.Options;
