+++ domains/auth/Auth.API/Tests/TestCryptographics.cs
@@ -1,4 +1,4 @@
-using API.Helpers;
+using API.Configuration;
 using API.Models;
 using API.Services;
 using Microsoft.Extensions.Options;
